move "wind_current_retrieval()" from cost_function.py to level2.py
