+++ seastar/retrieval/__init__.py
@@ -1,4 +1,4 @@
 import seastar.retrieval.level2
 import seastar.retrieval.auxiliary
 import seastar.retrieval.cost_function
-import seastar.retrieval.ambiguityremoval
+import seastar.retrieval.ambiguity_removal
